+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * `iterate` and `reverseIterate` now use a stack rather than an array
     internally.
   * Moved TypeScript types to the `types/` subdirectory.
+  * Setup ESLint and fixed lint issues.
 
 ## v0.10.0 (2025-06-22)